requirements.txt
python-dotenv
ttkbootstrap==1.10.1